fix corner cases in create and other fixes/enhancements
- check for "stopping" state of container during create when waiting
  for it before restart because that state can appear even before the
  entrypoint script's update to the status file becomes visible
- only trap HUP and TERM in entrypoint script for cleanup and ignore
  others (especially SIGPIPE which can be received in some cases)
- improved XAUTHORITY handling to deal with cases where one switches
  display manager resulting in totally different location; now handle
  three cases: 1) in /run/user/<uid> 2) in /tmp, 3) elsewhere
  for the first two cases only those prefixes are mounted in container
  while for the third case the parent directory is mounted like before
  (this handles cases where a subdir is being used for 1 or 2 that may
   disappear when switching display manager)
- improved default ybox apt conf to allow recommends/suggests as
  marked dependencies though not chosen by default in install
- create ~/.local/share/applications, ~/.local/bin dirs before copying
  wrappers there if the dirs do not exist
- added few more utilities to deb-generic default packages
- use Iterable instead of list in cmd.py:check_ybox_state and pass
  tuples instead of lists
- added more docstring comments in arch/pkgdeps.py to explain the
  rationale and process of building pacman/AUR maps of entire metadata

+++ src/ybox/conf/distros/deb-generic/init.sh
@@ -17,15 +17,16 @@ fi
 echo_color "$fg_cyan" "Configuring apt and updating package list" >> $status_file
 export HOME=/root
 export DEBIAN_FRONTEND=noninteractive
-# don't install recommended and suggested packages by default
+# don't install recommended and suggested packages by default but keep them if they are
+# deliberately marked as dependencies
 cat > /etc/apt/apt.conf.d/10-ybox << EOF
-APT::Get::Install-Recommends "0";
-APT::Get::Install-Suggests "0";
-APT::Install-Recommends "0";
-APT::Install-Suggests "0";
+APT::Install-Recommends "false";
+APT::Install-Suggests "false";
+APT::AutoRemove::RecommendsImportant "true";
+APT::AutoRemove::SuggestsImportant "true";
 EOF
 # allow language translations in apt for non english locales
-if [[ ! "$LANG" =~ ^en_.* ]]; then
+if [[ -n "$LANG" && ! "$LANG" =~ ^C(\..+)?$ && "$LANG" != "POSIX" && ! "$LANG" == en_* ]]; then
   rm -f /etc/apt/apt.conf.d/docker-no-languages
 fi
